Include labels when matching branches
Previously if the cursor was below the update-ref line, but still before
the label line, that branch would not be included in the consideration
for pushing the current branch. By including the label in the match, we
can now also capture these cases.

The additional filter is to make sure that every label we consider, also
has a corresponding update-ref line, otherwise we are in danger of
capturing labels like "onto".