Add integration for git-jump
The default installation of git offers a support script named git-jump
that outputs the locations of hunks in a quickfix-compatible format,
which can be used for example for the diff between revs, merge conflict
locations, and more [0].
We can use this script to navigate to such locations directly inside of
vim without having to call the script externally again.

Some example usecases for which hunks can be navigated with this:
- Unstaged changes with ":FirebaseJump diff"
- Staged and unstaged changes with ":FirebaseJump diff HEAD"
- Changes in the last commit with ":FirebaseJump diff HEAD~1 HEAD"

Some basic autocompletion support for the most common options is
provided, more advanced completion support such as git refs can be added
later.

For now we hardcode the location to the default path in /usr/share/git,
but this will not work with distros such as NixOS. Let's not wake up
sleeping dogs, this can be handled if there ever is any such user of
this niche feature on such a niche distro.

[0] https://github.com/git/git/tree/master/contrib/git-jump