Compute jump diff in correct direction
The diff needs to be inverted, as otherwise we look at a virtual revert
of the last commit.
This fixes line numbers being potentially out-of-sync with the real
file, with the only disadvantage being that now the entries of the
quickfix list can sometimes mismatch with the actual content, for
example in the case of deleted lines.

However, populating the quickfix list with the correct line locations is
the higher priority.